% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{create a new R project}
\usage{
create_project(path)
}
\arguments{
\item{path}{A path. If it does not exist, it is
created, provided that the parent path exists.}
}
\value{
path to the new package, invisibly
}
\description{
create a new R project
}
