% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{write_targets_script}
\alias{write_targets_script}
\title{initializes targets and creates minimal _targets.R template}
\usage{
write_targets_script()
}
\description{
initializes targets and creates minimal _targets.R template
}
